NameChange Resolution + Split Content files Per standards
